refactor tests into cli submodules
