+++ tests/cli/mod.rs
@@ -1,4 +1,4 @@
-// tests/cli.rs
+// tests/cli/mod.rs
 
 use assert_cmd::prelude::*;
 use assert_cmd::Command;
@@ -26,338 +26,13 @@ use users::{get_current_gid, get_current_uid, get_group_by_gid, get_user_by_uid}
 #[cfg(all(unix, feature = "xattr"))]
 use xattr as _;
 
-#[cfg(unix)]
-struct Tmpfs(TempDir);
+mod common;
+mod filtering;
+mod ownership;
+mod metadata;
 
 #[cfg(unix)]
-impl Tmpfs {
-    fn new() -> Option<Self> {
-        if !cfg!(target_os = "linux") {
-            return None;
-        }
-        if !Uid::effective().is_root() {
-            return None;
-        }
-        let mount_exists = std::env::var_os("PATH").is_some_and(|paths| {
-            std::env::split_paths(&paths).any(|dir| dir.join("mount").is_file())
-        });
-        if !mount_exists {
-            return None;
-        }
-        if let Ok(fs) = std::fs::read_to_string("/proc/filesystems") {
-            if !fs.lines().any(|l| l.trim().ends_with("tmpfs")) {
-                return None;
-            }
-        } else {
-            return None;
-        }
-        let dir = tempdir().ok()?;
-        let status = std::process::Command::new("mount")
-            .args(["-t", "tmpfs", "tmpfs", dir.path().to_str().unwrap()])
-            .stdout(std::process::Stdio::null())
-            .stderr(std::process::Stdio::null())
-            .status()
-            .ok()?;
-        if status.success() {
-            Some(Tmpfs(dir))
-        } else {
-            None
-        }
-    }
-
-    fn path(&self) -> &std::path::Path {
-        self.0.path()
-    }
-}
-
-#[test]
-fn files_from_from0_matches_rsync() {
-    use std::process::Command as StdCommand;
-
-    let tmp = tempdir().unwrap();
-    let src = tmp.path().join("src");
-    let rsync_dst = tmp.path().join("rsync");
-    let ours_dst = tmp.path().join("ours");
-    fs::create_dir_all(&src).unwrap();
-    fs::create_dir_all(&rsync_dst).unwrap();
-    fs::create_dir_all(&ours_dst).unwrap();
-
-    fs::write(src.join("include_me.txt"), "hi").unwrap();
-    fs::write(src.join("omit.log"), "nope").unwrap();
-
-    let list = tmp.path().join("list");
-    fs::write(&list, b"include_me.txt\0omit.log\0").unwrap();
-
-    let src_arg = format!("{}/", src.display());
-
-    StdCommand::new("rsync")
-        .args([
-            "-r",
-            "--from0",
-            "--files-from",
-            list.to_str().unwrap(),
-            "--exclude",
-            "*.log",
-            &src_arg,
-            rsync_dst.to_str().unwrap(),
-        ])
-        .status()
-        .unwrap();
-
-    Command::cargo_bin("oc-rsync")
-        .unwrap()
-        .args([
-            "--local",
-            "--recursive",
-            "--from0",
-            "--files-from",
-            list.to_str().unwrap(),
-            "--exclude",
-            "*.log",
-            &src_arg,
-            ours_dst.to_str().unwrap(),
-        ])
-        .assert()
-        .success();
-
-    let diff = StdCommand::new("diff")
-        .arg("-r")
-        .arg(&rsync_dst)
-        .arg(&ours_dst)
-        .status()
-        .unwrap();
-    assert!(diff.success(), "directory trees differ");
-}
-
-#[test]
-fn include_from_from0_matches_rsync() {
-    use std::process::Command as StdCommand;
-
-    let tmp = tempdir().unwrap();
-    let src = tmp.path().join("src");
-    let rsync_dst = tmp.path().join("rsync");
-    let ours_dst = tmp.path().join("ours");
-    fs::create_dir_all(&src).unwrap();
-    fs::create_dir_all(&rsync_dst).unwrap();
-    fs::create_dir_all(&ours_dst).unwrap();
-
-    fs::write(src.join("a.txt"), "hi").unwrap();
-    fs::write(src.join("b.log"), "nope").unwrap();
-    fs::write(src.join("c.txt"), "hi").unwrap();
-
-    let list = tmp.path().join("list");
-    fs::write(&list, b"a.txt\0c.txt\0").unwrap();
-
-    let src_arg = format!("{}/", src.display());
-
-    StdCommand::new("rsync")
-        .args([
-            "-r",
-            "--from0",
-            "--include-from",
-            list.to_str().unwrap(),
-            "--exclude",
-            "*",
-            &src_arg,
-            rsync_dst.to_str().unwrap(),
-        ])
-        .status()
-        .unwrap();
-
-    Command::cargo_bin("oc-rsync")
-        .unwrap()
-        .args([
-            "--local",
-            "--recursive",
-            "--from0",
-            "--include-from",
-            list.to_str().unwrap(),
-            "--exclude",
-            "*",
-            &src_arg,
-            ours_dst.to_str().unwrap(),
-        ])
-        .assert()
-        .success();
-
-    let diff = StdCommand::new("diff")
-        .arg("-r")
-        .arg(&rsync_dst)
-        .arg(&ours_dst)
-        .status()
-        .unwrap();
-    assert!(diff.success(), "directory trees differ");
-}
-
-#[test]
-fn exclude_from_from0_matches_rsync() {
-    use std::process::Command as StdCommand;
-
-    let tmp = tempdir().unwrap();
-    let src = tmp.path().join("src");
-    let rsync_dst = tmp.path().join("rsync");
-    let ours_dst = tmp.path().join("ours");
-    fs::create_dir_all(&src).unwrap();
-    fs::create_dir_all(&rsync_dst).unwrap();
-    fs::create_dir_all(&ours_dst).unwrap();
-
-    fs::write(src.join("keep.txt"), "hi").unwrap();
-    fs::write(src.join("drop.txt"), "nope").unwrap();
-
-    let list = tmp.path().join("list");
-    fs::write(&list, b"drop.txt\0").unwrap();
-
-    let src_arg = format!("{}/", src.display());
-
-    StdCommand::new("rsync")
-        .args([
-            "-r",
-            "--from0",
-            "--exclude-from",
-            list.to_str().unwrap(),
-            &src_arg,
-            rsync_dst.to_str().unwrap(),
-        ])
-        .status()
-        .unwrap();
-
-    Command::cargo_bin("oc-rsync")
-        .unwrap()
-        .args([
-            "--local",
-            "--recursive",
-            "--from0",
-            "--exclude-from",
-            list.to_str().unwrap(),
-            &src_arg,
-            ours_dst.to_str().unwrap(),
-        ])
-        .assert()
-        .success();
-
-    let diff = StdCommand::new("diff")
-        .arg("-r")
-        .arg(&rsync_dst)
-        .arg(&ours_dst)
-        .status()
-        .unwrap();
-    assert!(diff.success(), "directory trees differ");
-}
-
-#[test]
-fn filter_file_from0_matches_rsync() {
-    use std::process::Command as StdCommand;
-
-    let tmp = tempdir().unwrap();
-    let src = tmp.path().join("src");
-    let rsync_dst = tmp.path().join("rsync");
-    let ours_dst = tmp.path().join("ours");
-    fs::create_dir_all(&src).unwrap();
-    fs::create_dir_all(&rsync_dst).unwrap();
-    fs::create_dir_all(&ours_dst).unwrap();
-
-    fs::write(src.join("a.txt"), "hi").unwrap();
-    fs::write(src.join("b.log"), "no").unwrap();
-    fs::write(src.join("c.txt"), "hi").unwrap();
-
-    let filter = tmp.path().join("filters");
-    fs::write(&filter, b"+ *.txt\0- *\0").unwrap();
-
-    let src_arg = format!("{}/", src.display());
-
-    StdCommand::new("rsync")
-        .args([
-            "-r",
-            "--from0",
-            "--filter",
-            &format!("merge {}", filter.display()),
-            &src_arg,
-            rsync_dst.to_str().unwrap(),
-        ])
-        .status()
-        .unwrap();
-
-    Command::cargo_bin("oc-rsync")
-        .unwrap()
-        .args([
-            "--local",
-            "--recursive",
-            "--from0",
-            "--filter-file",
-            filter.to_str().unwrap(),
-            &src_arg,
-            ours_dst.to_str().unwrap(),
-        ])
-        .assert()
-        .success();
-
-    let diff = StdCommand::new("diff")
-        .arg("-r")
-        .arg(&rsync_dst)
-        .arg(&ours_dst)
-        .status()
-        .unwrap();
-    assert!(diff.success(), "directory trees differ");
-}
-
-#[test]
-fn per_dir_merge_matches_rsync() {
-    use std::process::Command as StdCommand;
-
-    let tmp = tempdir().unwrap();
-    let src = tmp.path().join("src");
-    let rsync_dst = tmp.path().join("rsync");
-    let ours_dst = tmp.path().join("ours");
-    fs::create_dir_all(&src).unwrap();
-    fs::create_dir_all(src.join("sub")).unwrap();
-    fs::create_dir_all(&rsync_dst).unwrap();
-    fs::create_dir_all(&ours_dst).unwrap();
-
-    fs::write(src.join("keep.txt"), "hi").unwrap();
-    fs::write(src.join("omit.log"), "no").unwrap();
-    fs::write(src.join("sub").join("keep2.txt"), "hi").unwrap();
-    fs::write(src.join("sub").join("omit2.txt"), "no").unwrap();
-
-    fs::write(src.join(".rsync-filter"), b"- *.log\n").unwrap();
-    fs::write(src.join("sub").join(".rsync-filter"), b"- omit2.txt\n").unwrap();
-
-    let src_arg = format!("{}/", src.display());
-
-    StdCommand::new("rsync")
-        .args(["-r", "-F", "-F", &src_arg, rsync_dst.to_str().unwrap()])
-        .status()
-        .unwrap();
-
-    Command::cargo_bin("oc-rsync")
-        .unwrap()
-        .args([
-            "--local",
-            "--recursive",
-            "-F",
-            "-F",
-            &src_arg,
-            ours_dst.to_str().unwrap(),
-        ])
-        .assert()
-        .success();
-
-    let diff = StdCommand::new("diff")
-        .arg("-r")
-        .arg(&rsync_dst)
-        .arg(&ours_dst)
-        .status()
-        .unwrap();
-    assert!(diff.success(), "directory trees differ");
-}
-
-#[cfg(unix)]
-impl Drop for Tmpfs {
-    fn drop(&mut self) {
-        let _ = std::process::Command::new("umount")
-            .arg(self.0.path())
-            .status();
-    }
-}
+use crate::common::Tmpfs;
 
 #[allow(clippy::vec_init_then_push)]
 #[test]
@@ -1560,140 +1235,6 @@ fn hard_links_preserved_via_cli() {
 
 #[cfg(unix)]
 #[test]
-fn numeric_ids_falls_back_when_unprivileged() {
-    let dir = tempdir().unwrap();
-    let probe = dir.path().join("probe");
-    std::fs::write(&probe, b"probe").unwrap();
-    let current_uid = get_current_uid();
-    let current_gid = get_current_gid();
-    if Uid::effective().is_root() {
-        eprintln!("skipping numeric_ids_falls_back_when_unprivileged: requires non-root");
-        return;
-    }
-    let target_uid = current_uid + 1;
-    if chown(&probe, Some(Uid::from_raw(target_uid)), None).is_ok() {
-        eprintln!("skipping numeric_ids_falls_back_when_unprivileged: has CAP_CHOWN");
-        return;
-    }
-
-    let src_dir = dir.path().join("src");
-    let dst_dir = dir.path().join("dst");
-    std::fs::create_dir_all(&src_dir).unwrap();
-    std::fs::create_dir_all(&dst_dir).unwrap();
-    let file = src_dir.join("id.txt");
-    std::fs::write(&file, b"ids").unwrap();
-
-    let src_arg = format!("{}/", src_dir.display());
-    Command::cargo_bin("oc-rsync")
-        .unwrap()
-        .args([
-            "--local",
-            "--numeric-ids",
-            "--owner",
-            "--group",
-            &src_arg,
-            dst_dir.to_str().unwrap(),
-        ])
-        .assert()
-        .success();
-
-    let dst_file = dst_dir.join("id.txt");
-    let meta = std::fs::metadata(&dst_file).unwrap();
-    assert_eq!(meta.uid(), current_uid);
-    assert_eq!(meta.gid(), current_gid);
-}
-
-#[cfg(unix)]
-#[test]
-fn owner_requires_privileges() {
-    let dir = tempdir().unwrap();
-    let probe = dir.path().join("probe");
-    std::fs::write(&probe, b"probe").unwrap();
-    let current_uid = get_current_uid();
-    let target_uid = if current_uid == 0 { 1 } else { current_uid + 1 };
-    if chown(&probe, Some(Uid::from_raw(target_uid)), None).is_ok() {
-        eprintln!("skipping owner_requires_privileges: has CAP_CHOWN or running as root");
-        return;
-    }
-
-    let src_dir = dir.path().join("src");
-    let dst_dir = dir.path().join("dst");
-    std::fs::create_dir_all(&src_dir).unwrap();
-    std::fs::create_dir_all(&dst_dir).unwrap();
-    let file = src_dir.join("id.txt");
-    std::fs::write(&file, b"ids").unwrap();
-
-    let src_arg = format!("{}/", src_dir.display());
-    Command::cargo_bin("oc-rsync")
-        .unwrap()
-        .args(["--local", "--owner", &src_arg, dst_dir.to_str().unwrap()])
-        .assert()
-        .failure()
-        .code(u8::from(protocol::ExitCode::StartClient) as i32)
-        .stderr(predicates::str::contains("changing ownership requires"));
-
-    let dst_file = dst_dir.join("id.txt");
-    assert!(!dst_file.exists());
-}
-
-#[cfg(unix)]
-#[test]
-fn user_and_group_ids_are_mapped() {
-    let uid = get_current_uid();
-    let _gid = get_current_gid();
-    if uid != 0 {
-        eprintln!("skipping user_and_group_ids_are_mapped: requires root or CAP_CHOWN");
-        return;
-    }
-    {
-        let dir = tempdir().unwrap();
-        let probe = dir.path().join("probe");
-        std::fs::write(&probe, b"probe").unwrap();
-        if let Err(err) = chown(&probe, Some(Uid::from_raw(1)), Some(Gid::from_raw(1))) {
-            match err {
-                nix::errno::Errno::EPERM => {
-                    eprintln!("skipping user_and_group_ids_are_mapped: lacks CAP_CHOWN");
-                    return;
-                }
-                _ => panic!("unexpected chown error: {err}"),
-            }
-        }
-    }
-
-    let dir = tempdir().unwrap();
-    let src_dir = dir.path().join("src");
-    let dst_dir = dir.path().join("dst");
-    std::fs::create_dir_all(&src_dir).unwrap();
-    std::fs::create_dir_all(&dst_dir).unwrap();
-    let file = src_dir.join("id.txt");
-    std::fs::write(&file, b"ids").unwrap();
-
-    let src_arg = format!("{}/", src_dir.display());
-    let uid = get_current_uid();
-    let gid = get_current_gid();
-    let mapped_uid = 1;
-    let mapped_gid = 1;
-    let usermap = format!("--usermap={uid}:{mapped_uid}");
-    let groupmap = format!("--groupmap={gid}:{mapped_gid}");
-    Command::cargo_bin("oc-rsync")
-        .unwrap()
-        .args([
-            "--local",
-            usermap.as_str(),
-            groupmap.as_str(),
-            src_arg.as_str(),
-            dst_dir.to_str().unwrap(),
-        ])
-        .assert()
-        .success();
-
-    let meta = std::fs::metadata(dst_dir.join("id.txt")).unwrap();
-    assert_eq!(meta.uid(), mapped_uid);
-    assert_eq!(meta.gid(), mapped_gid);
-}
-
-#[cfg(unix)]
-#[test]
 fn user_names_are_mapped_even_with_numeric_ids() {
     let uid = get_current_uid();
     if uid != 0 {
@@ -2284,76 +1825,6 @@ fn force_removes_multiple_non_empty_dirs() {
 #[cfg(unix)]
 #[test]
 #[serial]
-fn perms_flag_preserves_permissions() {
-    use nix::sys::stat::{umask, Mode};
-    use std::fs;
-    let dir = tempdir().unwrap();
-    let src_dir = dir.path().join("src");
-    let dst_dir = dir.path().join("dst");
-    fs::create_dir_all(&src_dir).unwrap();
-    fs::create_dir_all(&dst_dir).unwrap();
-    let file = src_dir.join("a.txt");
-    fs::write(&file, b"hi").unwrap();
-    fs::set_permissions(&file, fs::Permissions::from_mode(0o741)).unwrap();
-    let dst_file = dst_dir.join("a.txt");
-    fs::copy(&file, &dst_file).unwrap();
-    fs::set_permissions(&dst_file, fs::Permissions::from_mode(0o600)).unwrap();
-    let mtime = FileTime::from_last_modification_time(&fs::metadata(&file).unwrap());
-    set_file_mtime(&dst_file, mtime).unwrap();
-
-    let old_umask = umask(Mode::from_bits_truncate(0o077));
-
-    let mut cmd = Command::cargo_bin("oc-rsync").unwrap();
-    let src_arg = format!("{}/", src_dir.display());
-    cmd.args(["--local", "--perms", &src_arg, dst_dir.to_str().unwrap()]);
-    cmd.assert().success();
-
-    umask(old_umask);
-
-    let mode = fs::metadata(dst_dir.join("a.txt"))
-        .unwrap()
-        .permissions()
-        .mode();
-    assert_eq!(mode & 0o7777, 0o741);
-}
-
-#[cfg(unix)]
-#[test]
-#[serial]
-fn default_umask_masks_permissions() {
-    use nix::sys::stat::{umask, Mode};
-    use std::fs;
-    use std::os::unix::fs::PermissionsExt;
-    let dir = tempdir().unwrap();
-    let src_dir = dir.path().join("src");
-    let dst_dir = dir.path().join("dst");
-    fs::create_dir_all(&src_dir).unwrap();
-    fs::create_dir_all(&dst_dir).unwrap();
-    let file = src_dir.join("a.sh");
-    fs::write(&file, b"hi").unwrap();
-    fs::set_permissions(&file, fs::Permissions::from_mode(0o754)).unwrap();
-
-    let old_umask = umask(Mode::from_bits_truncate(0o027));
-
-    let src_arg = format!("{}/", src_dir.display());
-    Command::cargo_bin("oc-rsync")
-        .unwrap()
-        .args(["--local", &src_arg, dst_dir.to_str().unwrap()])
-        .assert()
-        .success();
-
-    umask(old_umask);
-
-    let mode = fs::metadata(dst_dir.join("a.sh"))
-        .unwrap()
-        .permissions()
-        .mode()
-        & 0o777;
-    assert_eq!(mode, 0o754 & !0o027);
-}
-
-#[cfg(unix)]
-#[test]
 fn chmod_masks_file_type_bits() {
     use std::fs;
     use std::os::unix::fs::PermissionsExt;
@@ -3551,56 +3022,6 @@ fn safe_links_allow_relative_symlink_within_tree() {
     assert!(meta.file_type().is_symlink());
     let target = std::fs::read_link(dst.join("dir/link")).unwrap();
     assert_eq!(target, std::path::PathBuf::from("../file"));
-}
-
-#[cfg(unix)]
-#[test]
-fn perms_preserve_permissions() {
-    let dir = tempdir().unwrap();
-    let src = dir.path().join("src");
-    let dst = dir.path().join("dst");
-    std::fs::create_dir_all(&src).unwrap();
-    let file = src.join("file");
-    std::fs::write(&file, b"hi").unwrap();
-    std::fs::set_permissions(&file, std::fs::Permissions::from_mode(0o640)).unwrap();
-
-    let src_arg = format!("{}/", src.display());
-    Command::cargo_bin("oc-rsync")
-        .unwrap()
-        .args(["--local", "--perms", &src_arg, dst.to_str().unwrap()])
-        .assert()
-        .success();
-
-    let meta = std::fs::metadata(dst.join("file")).unwrap();
-    assert_eq!(meta.permissions().mode() & 0o7777, 0o640);
-}
-
-#[cfg(unix)]
-#[test]
-fn times_preserve_mtime() {
-    let dir = tempdir().unwrap();
-    let src = dir.path().join("src");
-    let dst = dir.path().join("dst");
-    std::fs::create_dir_all(&src).unwrap();
-    std::fs::create_dir_all(&dst).unwrap();
-    let file = src.join("file");
-    std::fs::write(&file, b"hi").unwrap();
-    let mtime = FileTime::from_unix_time(1_000_000, 0);
-    set_file_mtime(&file, mtime).unwrap();
-    let dst_file = dst.join("file");
-    std::fs::copy(&file, &dst_file).unwrap();
-    set_file_mtime(&dst_file, FileTime::from_unix_time(0, 0)).unwrap();
-
-    let src_arg = format!("{}/", src.display());
-    Command::cargo_bin("oc-rsync")
-        .unwrap()
-        .args(["--local", "--times", &src_arg, dst.to_str().unwrap()])
-        .assert()
-        .success();
-
-    let meta = std::fs::metadata(dst.join("file")).unwrap();
-    let dst_mtime = FileTime::from_last_modification_time(&meta);
-    assert_eq!(dst_mtime, mtime);
 }
 
 #[cfg(unix)]
